Add support to specific route in express (#18)
A new option was added that allows to use a specific route
with the middleware provided by staticify.

This option is used by the `getVersionedPath` method.

The specified route for the middleware is added before the
versioned path. When a specific route is not used, the method
continues to function normally.